+++ src/components/SearchResult.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="max-w-sm mx-auto flex flex-col items-center justify-center">
+    <h1>resultado</h1>
+    <h2>{{ resultado }}</h2>
+  </div>
+</template>
+
+<script>
+export default {
+  props: ['resultado'],
+  data() {
+    return {
+    };
+  },
+  methods: {
+  },
+};
+</script>
